Cloud Functions/Twilio - Daily SMS/requirements.txt
twilio
google-cloud-bigquery==2.34.3
pandas
